requirements.txt
aiohttp==3.4.1
async-timeout==3.0.0
attrs==18.1.0
bencoder.pyx==1.2.1
chardet==3.0.4
idna==2.7
Maga==3.0.0
multidict==4.3.1
mysql-connector-python==8.0.12
protobuf==3.6.1
pytz==2018.5
redis==2.10.6
six==1.11.0
yarl==1.2.6
